Moved admin analytics folders to kebab-case names
no ref

- apps/admin/src uses kebab-case directory names everywhere except the
  analytics family, which kept PascalCase folders from its former repo
- lower-cased src/analytics/views/Stats (+ Growth, Locations, Newsletters,
  Overview, Web) and src/posts/analytics/{Growth,Newsletter,Overview,Web}
- mechanical rename only: directory moves plus matching import path updates
