+++ apps/admin/src/analytics/views/stats/components/feature-image-placeholder.tsx
