+++ apps/admin/src/analytics/views/stats/growth/growth.tsx
@@ -1,14 +1,14 @@
-import DateRangeSelect from '@/analytics/views/Stats/components/date-range-select';
+import DateRangeSelect from '@/analytics/views/stats/components/date-range-select';
 import DisabledSourcesIndicator from '@/shared/analytics/disabled-sources-indicator';
 import GrowthKPIs from './components/growth-kpis';
 import GrowthSources from './components/growth-sources';
 import NewSubscribersCadence from './components/new-subscribers-cadence';
 import PaidMembersChangeChart from './components/paid-subscription-change-chart';
 import React, {useEffect, useMemo, useState} from 'react';
-import SortButton from '@/analytics/views/Stats/components/sort-button';
-import StatsHeader from '@/analytics/views/Stats/layout/stats-header';
-import StatsLayout from '@/analytics/views/Stats/layout/stats-layout';
-import StatsView from '@/analytics/views/Stats/layout/stats-view';
+import SortButton from '@/analytics/views/stats/components/sort-button';
+import StatsHeader from '@/analytics/views/stats/layout/stats-header';
+import StatsLayout from '@/analytics/views/stats/layout/stats-layout';
+import StatsView from '@/analytics/views/stats/layout/stats-view';
 import {Button, Card, CardContent, CardDescription, CardHeader, CardTitle, EmptyIndicator, NavbarActions, Skeleton, Table, TableBody, TableCell, TableHead, TableHeader, TableRow, Tabs, TabsList, TabsTrigger} from '@tryghost/shade/components';
 import {CONTENT_TYPES, type ContentType, getContentTitle, getGrowthContentDescription} from '@/analytics/utils/content-helpers';
 import {LucideIcon, formatDisplayDate, formatNumber} from '@tryghost/shade/utils';
