+++ apps/admin/src/analytics/views/stats/layout/stats-content.tsx
