+++ apps/admin/src/analytics/views/stats/layout/stats-view.tsx
