+++ apps/admin/src/analytics/views/stats/locations/components/locations-card.tsx
