+++ apps/admin/src/analytics/views/stats/newsletters/index.ts
