+++ apps/admin/src/analytics/views/stats/web/index.ts
