+++ apps/admin/src/posts/analytics/web/components/locations.tsx
